update 003 20/11/2024

+++ src/main/java/com/nts/awspremium/model/Service.java
@@ -86,7 +86,7 @@ public class Service implements Serializable {
     @Column(columnDefinition = "integer default 0")
     private Integer youtube_reply;
     @Column(columnDefinition = "integer default 0")
-    private Float bonus ;
+    private Integer bonus ;
     @Column(columnDefinition = "integer default 0")
     private Integer check_done ;
     @Column(columnDefinition = "integer default 0")
